version with ticket storage
